Split test_safe_cmd_run.py into cohesive modules (#117)
test_safe_cmd_run.py was 828 lines, over the 400-line CodeScene "Lines of Code
in a Single File" limit. Move tests into focused modules by concern, without
changing any test behaviour, assertions, names, or async/sync coverage (the
collected node-id set is unchanged):

- test_safe_cmd_stdin.py (extended): stdin injection tests (text/bytes feeding,
  configured encoding, capture-disabled, early-close, and the forbidden-command
  vs stdin-encoding ordering contract) alongside the existing stdin writer
  lifecycle regressions. Adds a local `execution_strategy` fixture.
- test_safe_cmd_context.py (new): allowlist enforcement and before/after hook
  integration (FIFO/LIFO order, hook arguments, cancellation skipping after
  hooks).
- test_safe_cmd_streams.py (new): captured-stream (`capture=True`) cleanup on
  cancellation and stdin-writer failure, preserving the monkeypatch targets.
- test_safe_cmd_run.py (slimmed to 366 lines): general execution/output/env/
  cwd/timeout coverage plus the non-cooperative-kill escalation test.

Regenerate the native-wheel snapshot to package the two new test modules. No
production code changed; no CodeScene suppression added.
