+++ cuprum/unittests/test_safe_cmd_run.py
@@ -13,18 +13,16 @@ from pathlib import Path
 
 import pytest
 
-from cuprum import ECHO, ForbiddenProgramError, ScopeConfig, TimeoutExpired, scoped, sh
+from cuprum import ECHO, TimeoutExpired, sh
 from cuprum.sh import (
     CommandResult,
     ExecutionContext,
     IOOptions,
     RunOutputOptions,
-    StdinInput,
 )
 from tests.helpers.catalogue import python_builder as build_python_builder
 
 if typ.TYPE_CHECKING:
-    from cuprum.events import ExecEvent
     from cuprum.sh import SafeCmd
 
 type ExecuteFn = cabc.Callable[[SafeCmd, dict[str, typ.Any]], CommandResult]
@@ -292,134 +290,6 @@ def test_decodes_with_configured_encoding(
     assert result.stderr == ""
 
 
-def test_input_text_feeds_stdin(
-    python_builder: cabc.Callable[..., SafeCmd],
-    execution_strategy: tuple[str, ExecuteFn],
-) -> None:
-    """Both run() and run_sync() feed text directly to stdin."""
-    _, execute = execution_strategy
-    command = python_builder("-c", "import sys; print(sys.stdin.read(), end='')")
-
-    result = execute(command, {"stdin": StdinInput(text="hello stdin\n")})
-
-    assert result.exit_code == 0
-    assert result.stdout == "hello stdin\n"
-    assert result.stderr == ""
-
-
-def test_input_bytes_feeds_raw_stdin(
-    python_builder: cabc.Callable[..., SafeCmd],
-    execution_strategy: tuple[str, ExecuteFn],
-) -> None:
-    """Both run() and run_sync() feed bytes directly to stdin."""
-    _, execute = execution_strategy
-    command = python_builder(
-        "-c",
-        "import sys; sys.stdout.buffer.write(sys.stdin.buffer.read())",
-    )
-
-    result = execute(command, {"stdin": StdinInput(data=b"\x00raw\xff\n")})
-
-    assert result.exit_code == 0
-    assert result.stdout == "\x00raw\ufffd\n"
-    assert result.stderr == ""
-
-
-def test_input_text_uses_configured_encoding(
-    python_builder: cabc.Callable[..., SafeCmd],
-    execution_strategy: tuple[str, ExecuteFn],
-) -> None:
-    """Text stdin is encoded using the execution context settings."""
-    _, execute = execution_strategy
-    command = python_builder(
-        "-c",
-        "import sys; sys.stdout.buffer.write(sys.stdin.buffer.read())",
-    )
-
-    result = execute(
-        command,
-        {
-            "stdin": StdinInput(text="\u2013"),
-            "context": ExecutionContext(encoding="cp1252", errors="strict"),
-        },
-    )
-
-    assert result.exit_code == 0
-    assert result.stdout == "\u2013"
-    assert result.stderr == ""
-
-
-def test_input_text_and_input_bytes_conflict(
-    python_builder: cabc.Callable[..., SafeCmd],
-    execution_strategy: tuple[str, ExecuteFn],
-) -> None:
-    """Supplying both stdin forms raises a validation error."""
-    _ = python_builder, execution_strategy
-    with pytest.raises(
-        ValueError,
-        match=r"text and data cannot both be provided",
-    ):
-        StdinInput(text="hello", data=b"hello")
-
-
-def test_input_text_works_when_capture_is_disabled(
-    python_builder: cabc.Callable[..., SafeCmd],
-    execution_strategy: tuple[str, ExecuteFn],
-) -> None:
-    """Stdin injection still works when stdout/stderr are not captured."""
-    _, execute = execution_strategy
-    command = python_builder(
-        "-c",
-        "import sys; sys.exit(0 if sys.stdin.read() == 'uncaptured' else 9)",
-    )
-
-    result = execute(
-        command,
-        {
-            "output": RunOutputOptions(capture=False),
-            "stdin": StdinInput(text="uncaptured"),
-        },
-    )
-
-    assert result.exit_code == 0
-    assert result.stdout is None
-    assert result.stderr is None
-
-
-def test_nonzero_exit_code_is_captured_with_input_text(
-    python_builder: cabc.Callable[..., SafeCmd],
-    execution_strategy: tuple[str, ExecuteFn],
-) -> None:
-    """Non-zero exits still include captured streams when stdin is supplied."""
-    _, execute = execution_strategy
-    command = python_builder(
-        "-c",
-        ("import sys; data = sys.stdin.read(); print(data, end=''); sys.exit(7)"),
-    )
-
-    result = execute(command, {"stdin": StdinInput(text="failure input")})
-
-    assert result.exit_code == 7
-    assert result.ok is False
-    assert result.stdout == "failure input"
-    assert result.stderr == ""
-
-
-def test_process_closing_stdin_early_is_handled(
-    python_builder: cabc.Callable[..., SafeCmd],
-    execution_strategy: tuple[str, ExecuteFn],
-) -> None:
-    """A process that ignores stdin does not fail the command run."""
-    _, execute = execution_strategy
-    command = python_builder("-c", "print('done')")
-
-    result = execute(command, {"stdin": StdinInput(text="ignored stdin")})
-
-    assert result.exit_code == 0
-    assert result.stdout == "done\n"
-    assert result.stderr == ""
-
-
 # -----------------------------------------------------------------------------
 # Async-only tests (cancellation semantics)
 # -----------------------------------------------------------------------------
@@ -494,335 +364,3 @@ def _poll_process_death(pid: int, *, timeout: float = 1.0) -> None:
             return
         time.sleep(0.02)
     pytest.fail(f"Process {pid} still alive after {timeout}s of polling")
-
-
-# -----------------------------------------------------------------------------
-# Context integration tests (allowlist and hooks)
-# -----------------------------------------------------------------------------
-
-
-def test_run_raises_forbidden_when_program_not_in_allowlist(
-    execution_strategy: tuple[str, ExecuteFn],
-) -> None:
-    """run() raises ForbiddenProgramError when program is not in context allowlist."""
-    from cuprum.program import Program
-
-    _, execute = execution_strategy
-    command = sh.make(ECHO)("hello")
-    # Create a context with an allowlist that does NOT include ECHO
-    other_program = Program("cat")
-    with (
-        scoped(ScopeConfig(allowlist=frozenset([other_program]))),
-        pytest.raises(ForbiddenProgramError, match=r"echo"),
-    ):
-        execute(command, {})
-
-
-def test_run_succeeds_when_program_in_allowlist(
-    execution_strategy: tuple[str, ExecuteFn],
-) -> None:
-    """run() succeeds when program is in context allowlist."""
-    from cuprum.context import ScopeConfig, scoped
-
-    _, execute = execution_strategy
-    command = sh.make(ECHO)("-n", "allowed")
-    with scoped(ScopeConfig(allowlist=frozenset([ECHO]))):
-        result = execute(command, {})
-    assert result.exit_code == 0
-    assert result.stdout == "allowed"
-
-
-def test_run_succeeds_with_empty_allowlist() -> None:
-    """run() succeeds when context allowlist is empty (default permits all)."""
-    # Default context has empty allowlist which permits all programs
-    command = sh.make(ECHO)("-n", "hello")
-    result = asyncio.run(command.run())
-    assert result.exit_code == 0
-    assert result.stdout == "hello"
-
-
-def test_run_invokes_before_hooks_in_fifo_order(
-    execution_strategy: tuple[str, ExecuteFn],
-) -> None:
-    """run() invokes before hooks in registration order (FIFO)."""
-    from cuprum.context import ScopeConfig, scoped
-
-    _, execute = execution_strategy
-    call_order: list[int] = []
-
-    def hook1(cmd: SafeCmd) -> None:
-        """Record this before hook as the first to run."""
-        _ = cmd
-        call_order.append(1)
-
-    def hook2(cmd: SafeCmd) -> None:
-        """Record this before hook as the second to run."""
-        _ = cmd
-        call_order.append(2)
-
-    command = sh.make(ECHO)("-n", "hooks")
-    with scoped(ScopeConfig(allowlist=frozenset([ECHO]), before_hooks=(hook1, hook2))):
-        execute(command, {})
-
-    assert call_order == [1, 2]
-
-
-def test_run_invokes_after_hooks_in_lifo_order(
-    execution_strategy: tuple[str, ExecuteFn],
-) -> None:
-    """run() invokes after hooks in LIFO order (inner hooks run before outer)."""
-    from cuprum.context import ScopeConfig, scoped
-
-    _, execute = execution_strategy
-    call_order: list[int] = []
-
-    def outer_hook(cmd: SafeCmd, result: CommandResult) -> None:
-        """Record the outer scope's after hook invocation."""
-        _, _ = cmd, result
-        call_order.append(1)
-
-    def inner_hook(cmd: SafeCmd, result: CommandResult) -> None:
-        """Record the inner scope's after hook invocation."""
-        _, _ = cmd, result
-        call_order.append(2)
-
-    command = sh.make(ECHO)("-n", "hooks")
-    # Nest scopes so the inner after hook runs before the outer (LIFO)
-    with scoped(  # noqa: SIM117 — nested scopes required for LIFO hook ordering test
-        ScopeConfig(allowlist=frozenset([ECHO]), after_hooks=(outer_hook,))
-    ):
-        with scoped(ScopeConfig(after_hooks=(inner_hook,))):
-            execute(command, {})
-
-    # Inner hook (2) runs first, then outer hook (1) - true LIFO semantics
-    assert call_order == [2, 1]
-
-
-def test_run_passes_command_and_result_to_hooks(
-    execution_strategy: tuple[str, ExecuteFn],
-) -> None:
-    """run() passes SafeCmd to before hooks and SafeCmd+result to after hooks."""
-    from cuprum.context import ScopeConfig, scoped
-
-    _, execute = execution_strategy
-    before_received: list[SafeCmd] = []
-    after_received: list[tuple[SafeCmd, CommandResult]] = []
-
-    def before_hook(cmd: SafeCmd) -> None:
-        """Capture the command passed to the before hook."""
-        before_received.append(cmd)
-
-    def after_hook(cmd: SafeCmd, result: CommandResult) -> None:
-        """Capture the command and result passed to the after hook."""
-        after_received.append((cmd, result))
-
-    command = sh.make(ECHO)("-n", "test")
-    with scoped(
-        ScopeConfig(
-            allowlist=frozenset([ECHO]),
-            before_hooks=(before_hook,),
-            after_hooks=(after_hook,),
-        )
-    ):
-        result = execute(command, {})
-
-    assert len(before_received) == 1
-    assert before_received[0] is command
-    assert len(after_received) == 1
-    assert after_received[0][0] is command
-    assert after_received[0][1] is result
-
-
-def test_run_does_not_invoke_after_hooks_on_cancellation(
-    python_builder: cabc.Callable[..., SafeCmd],
-) -> None:
-    """run() does not invoke after hooks when task is cancelled."""
-    from cuprum.context import ScopeConfig, scoped
-
-    after_called = False
-
-    def after_hook(cmd: SafeCmd, result: CommandResult) -> None:
-        """Record that the after hook was invoked."""
-        nonlocal after_called
-        _, _ = cmd, result
-        after_called = True
-
-    # Use a long-running command that we can cancel
-    command = python_builder("-c", "import time; time.sleep(10)")
-
-    async def orchestrate() -> None:
-        """Start the command under a scope, then cancel it."""
-        with scoped(
-            ScopeConfig(
-                allowlist=frozenset([command.program]), after_hooks=(after_hook,)
-            )
-        ):
-            task = asyncio.create_task(
-                command.run(
-                    output=RunOutputOptions(capture=False),
-                    context=ExecutionContext(cancel_grace=0.1),
-                ),
-            )
-            await asyncio.sleep(0.1)  # Let the process start
-            task.cancel()
-            with pytest.raises(asyncio.CancelledError):
-                await task
-
-    asyncio.run(orchestrate())
-    assert after_called is False
-
-
-def test_streamed_run_cancellation_cleans_up_task(
-    python_builder: cabc.Callable[..., SafeCmd],
-) -> None:
-    """Cancelling an in-flight streamed (capture=True) run does not hang.
-
-    Output capture routes execution through ``_run_subprocess_with_streams``,
-    so cancelling the run must cancel and drain the stdout/stderr consumer
-    tasks via ``_wait_for_exit_code``'s ``CancelledError`` branch rather than
-    deadlocking on a still-pending reader (regression for #117).
-    """
-
-    async def _orchestrate() -> None:
-        """Start a streaming command and cancel it once output is observed."""
-        command = python_builder(
-            "-c",
-            "import time; [print('tick', flush=True) or time.sleep(0.02) "
-            "for _ in range(600)]",
-        )
-        streaming = asyncio.Event()
-
-        def on_event(ev: ExecEvent) -> None:
-            """Signal once a stdout consumer has delivered a line."""
-            if ev.phase == "stdout":
-                streaming.set()
-
-        with (
-            scoped(ScopeConfig(allowlist=frozenset([command.program]))),
-            sh.observe(on_event),
-        ):
-            task = asyncio.create_task(
-                command.run(output=RunOutputOptions(capture=True)),
-            )
-            # Cancel only after a stream consumer has actually delivered a
-            # line: a deterministic readiness signal rather than a fixed sleep.
-            await asyncio.wait_for(streaming.wait(), timeout=2.0)
-            task.cancel()
-            with pytest.raises(asyncio.CancelledError):
-                await asyncio.wait_for(asyncio.shield(task), timeout=2.0)
-            assert task.cancelled(), (
-                "cancellation must propagate out of the streamed run; the task "
-                "must not swallow CancelledError"
-            )
-
-    asyncio.run(_orchestrate())
-
-
-def test_streamed_run_reconciles_consumers_on_stdin_writer_failure(
-    python_builder: cabc.Callable[..., SafeCmd],
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """A failing stdin writer must not orphan the stream consumer tasks.
-
-    On the streamed (capture) path, an unexpected exception raised from
-    ``await stdin_task`` must still cancel and drain the stdout/stderr consumer
-    tasks before it propagates, rather than leaving them pending (regression
-    for #117).
-    """
-
-    class _InjectedStdinError(Exception):
-        """Sentinel error injected by the fake stdin writer."""
-
-    recorded: list[asyncio.Task[str | None]] = []
-
-    async def _raise_stdin(
-        process: asyncio.subprocess.Process,
-        stdin_data: bytes,
-        observation: object,
-    ) -> None:
-        """Close stdin, then fail the writer with an injected error."""
-        _ = (stdin_data, observation)
-        await asyncio.sleep(0)  # yield like the real writer's drain() await
-        if process.stdin is not None:
-            process.stdin.close()
-        raise _InjectedStdinError
-
-    def _spawn_blocking_consumers(
-        process: object,
-        execution: object,
-        stream_config: object,
-        *,
-        pid: int | None,
-    ) -> tuple[asyncio.Task[str | None], asyncio.Task[str | None]]:
-        """Return two never-completing consumer tasks and record them."""
-        _ = (process, execution, stream_config, pid)
-
-        async def _block() -> str | None:
-            """Block until cancelled during stdin-failure cleanup."""
-            await asyncio.Event().wait()
-
-        tasks = (asyncio.create_task(_block()), asyncio.create_task(_block()))
-        recorded.extend(tasks)
-        return tasks
-
-    monkeypatch.setattr("cuprum._subprocess_stdin._write_stdin", _raise_stdin)
-    monkeypatch.setattr(
-        "cuprum._subprocess_execution._spawn_stream_consumers",
-        _spawn_blocking_consumers,
-    )
-
-    command = python_builder("-c", "pass")
-
-    async def _drive() -> None:
-        """Run the streamed command and inspect consumers within the loop."""
-        with pytest.raises(_InjectedStdinError):
-            await command.run(
-                output=RunOutputOptions(capture=True),
-                stdin=StdinInput(data=b"payload"),
-            )
-        # Assert inside the still-running loop, before asyncio.run tears it
-        # down: loop teardown would itself cancel any orphaned tasks and mask
-        # the bug, so only the reconcile path can have finished these here.
-        assert len(recorded) == 2, "streamed path must spawn two stream consumers"
-        assert all(task.cancelled() for task in recorded), (
-            "stdout/stderr consumers must be cancelled and drained when the "
-            "stdin writer fails, not left orphaned as pending tasks"
-        )
-
-    asyncio.run(_drive())
-
-
-@pytest.mark.parametrize("execution_strategy", ["async", "sync"])
-def test_input_text_encoding_failure_raises(
-    python_builder: cabc.Callable[..., SafeCmd],
-    execution_strategy: str,
-) -> None:
-    """UnicodeEncodeError propagates when text cannot be encoded."""
-    command = python_builder("-c", "import sys; sys.stdin.read()")
-    ctx = ExecutionContext(encoding="ascii", errors="strict")
-    execute = _execute_async if execution_strategy == "async" else _execute_sync
-    with pytest.raises(UnicodeEncodeError):
-        execute(
-            command,
-            {"stdin": StdinInput(text="\u00e9 non-ASCII"), "context": ctx},
-        )
-
-
-@pytest.mark.parametrize("execution_strategy", ["async", "sync"])
-def test_forbidden_command_rejects_before_stdin_encoding(
-    python_builder: cabc.Callable[..., SafeCmd],
-    execution_strategy: str,
-) -> None:
-    """Forbidden commands fail before stdin encoding is attempted."""
-    command = python_builder("-c", "import sys; sys.stdin.read()")
-    ctx = ExecutionContext(encoding="ascii", errors="strict")
-    execute = _execute_async if execution_strategy == "async" else _execute_sync
-
-    with (
-        scoped(ScopeConfig(allowlist=frozenset([ECHO]))),
-        pytest.raises(ForbiddenProgramError),
-    ):
-        execute(
-            command,
-            {"stdin": StdinInput(text="\u00e9 non-ASCII"), "context": ctx},
-        )
